fix: achievements position
