+++ frontend/src/pages/Profile.vue
@@ -140,7 +140,7 @@ const coverImage = createResource({
 
 const setActiveTab = () => {
 	let fragments = route.path.split('/')
-	let sections = ['certificates', 'achievements', 'roles', 'evaluations']
+	let sections = ['certificates', 'roles', 'evaluations']
 	sections.forEach((section) => {
 		if (fragments.includes(section)) {
 			activeTab.value = convertToTitleCase(section)
@@ -154,7 +154,6 @@ watchEffect(() => {
 		let route = {
 			About: { name: 'ProfileAbout' },
 			Certificates: { name: 'ProfileCertificates' },
-			Achievements: { name: 'ProfileAchievements' },
 			Roles: { name: 'ProfileRoles' },
 			Evaluations: { name: 'ProfileEvaluator' },
 		}[activeTab.value]
@@ -171,11 +170,7 @@ const isSessionUser = () => {
 }
 
 const getTabButtons = () => {
-	let buttons = [
-		{ label: 'About' },
-		{ label: 'Certificates' },
-		{ label: 'Achievements' },
-	]
+	let buttons = [{ label: 'About' }, { label: 'Certificates' }]
 	if ($user.data?.is_moderator) buttons.push({ label: 'Roles' })
 	if (isSessionUser() && $user.data?.is_evaluator)
 		buttons.push({ label: 'Evaluations' })
