+++ frontend/src/router.js
@@ -73,11 +73,6 @@ const routes = [
 				component: () => import('@/pages/ProfileCertificates.vue'),
 			},
 			{
-				name: 'ProfileAchievements',
-				path: 'achievements',
-				component: () => import('@/pages/ProfileAchievements.vue'),
-			},
-			{
 				name: 'ProfileRoles',
 				path: 'roles',
 				component: () => import('@/pages/ProfileRoles.vue'),
