+++ lms/lms/doctype/lms_course/lms_course.json
@@ -274,7 +274,7 @@
   }
  ],
  "make_attachments_public": 1,
- "modified": "2024-04-16 17:40:50.899368",
+ "modified": "2024-05-08 15:11:07.833094",
  "modified_by": "Administrator",
  "module": "LMS",
  "name": "LMS Course",
@@ -305,7 +305,7 @@
    "write": 1
   }
  ],
- "search_fields": "title",
+ "search_fields": "title, tags",
  "show_title_field_in_link": 1,
  "sort_field": "creation",
  "sort_order": "DESC",
